Pull base currency from website and not global level
Follwing the pull request #28 from @pevik to make the extension multistore-ready, the currency code needs to be pulled from the website level, not global level. 
If you have different base currencies on global and website level, a website specific base currency setting will be overwritten by the global level and the value converted to the checkout currency from the global level currency, not the website currency. 

This pull request fixes this issue and pulls the website level base currency.